+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Unreleased
 ==========
 * Dropped Support for Python 3.8
 * Introduced Django CMS 4.1 support.
+* delete `VersionLock` as It's moved to Version.locked_by field.
 
 
 1.3.0 (2024-05-16)